Partial batch job results repository implementation
