+++ webpack.config.js
@@ -61,6 +61,7 @@ module.exports = {
       __OWLEYES_ENDPOINT__: `'${process.env.OWLEYES_ENDPOINT}'`,
       __SEENOMALY_ENDPOINT__: `'${process.env.SEENOMALY_ENDPOINT}'`,
       __BATCH_JOB_ENDPOINT__: `'${process.env.BATCH_JOB_ENDPOINT}'`,
+      __BATCH_JOB_REPORT_ENDPOINT__: `'${process.env.BATCH_JOB_REPORT_ENDPOINT}'`,
     }),
   ],
 };
